[spark] Merge partition predicate tests into SparkPredicateUtilsTest
The new SparkPartitionPredicate tests and the existing
SparkPredicateConverter tests are both unit tests for predicate utilities
in the same package and share the same Spark DSv2 predicate helpers.
Consolidate them into a single SparkPredicateUtilsTest suite so that
ref/lit/pred helpers are not duplicated and the test surface is easier to
maintain.
